backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Pydantic
pydantic==2.6.0
pydantic-settings==2.2.0

# OpenAI
openai==1.12.0

# Embeddings
sentence-transformers==2.3.1
torch>=2.0.0

# Document parsing
pypdf==4.0.1
cryptography>=3.1  # Required for encrypted PDFs
python-docx==1.1.2
markdown==3.6

# Token counting
tiktoken==0.7.0

# Task queue
celery==5.3.6
redis==5.0.6

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1

